maps functionalitites (working frontend)

+++ app/src/main/java/com/krishna/mutemate/model/LocationMute.kt
@@ -13,6 +13,6 @@ data class LocationMute(
     @DrawableRes
     val markerType: Int,
     val muteOptions: AllMuteOptions,
-    val radius: Int? = null,
+    val radius: Float? = null,
     val latLng: LatLng? = null,
 )
